requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.5
django-filter==23.2
Pillow==10.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
whitenoise==6.5.0
